# Apps/StubCustodians/src/SUI.Custodians.API.Client/custodians-openapi-generated.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Custodians API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Find/{orgId}/{personId}": {
      "get": {
        "tags": [
          "Find"
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "personId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          },
          {
            "name": "recordType",
            "in": "query",
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  },
                  "additionalProperties": false
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  },
                  "additionalProperties": false
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Find/manifest": {
      "post": {
        "tags": [
          "Find"
        ],
        "requestBody": {
          "content": {
            "application/json; x-api-version=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ManifestRequest"
              }
            },
            "text/json; x-api-version=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ManifestRequest"
              }
            },
            "application/*+json; x-api-version=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  },
                  "additionalProperties": false
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  },
                  "additionalProperties": false
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultItem"
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "application/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              },
              "text/json; x-api-version=1.0": {
                "schema": {
                  "$ref": "#/components/schemas/FailureInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorInfo": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "message": {
            "type": "string",
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "FailureInfo": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/FailureKind"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
      "FailureKind": {
        "type": "integer",
        "additionalProperties": false
      },
...
      "ManifestRequest": {
        "required": [
          "orgId",
          "personId",
          "recordType"
        ],
        "type": "object",
        "properties": {
          "orgId": {
            "type": "string",
            "additionalProperties": false
          },
          "personId": {
            "type": "string",
            "additionalProperties": false
          },
          "recordType": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "title": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "status": {
            "type": "integer",
            "additionalProperties": false,
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          },
          "instance": {
            "type": "string",
            "additionalProperties": false,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SearchResultItem": {
        "required": [
          "providerSystem",
          "providerName",
          "recordType",
          "recordUrl"
        ],
        "type": "object",
        "properties": {
          "providerSystem": {
            "type": "string",
            "additionalProperties": false
          },
          "providerName": {
            "type": "string",
            "additionalProperties": false
          },
          "recordType": {
            "type": "string",
            "additionalProperties": false
          },
          "recordUrl": {
            "type": "string",
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}